+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/models/RedactReport.java
@@ -1,22 +1,20 @@
-package cloud.pangeacyber.pangea.data_guard.models;
+package cloud.pangeacyber.pangea.ai_guard.models;
 
-import cloud.pangeacyber.pangea.intel.models.UserBreachedData;
+import cloud.pangeacyber.pangea.redact.models.RedactRecognizerResult;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import java.util.List;
 import lombok.Value;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(Include.NON_NULL)
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 @Value
-public final class TextGuardReport {
+public final class RedactReport {
 
-	IntelResults domainIntel;
-	IntelResults ipIntel;
-	RedactReport redact;
-	IntelResults urlIntel;
-	UserBreachedData userIntel;
+	int count;
+	List<RedactRecognizerResult> recognizerResults;
 }
