+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/models/TextGuardArtifact.java
@@ -1,4 +1,4 @@
-package cloud.pangeacyber.pangea.data_guard.models;
+package cloud.pangeacyber.pangea.ai_guard.models;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
