+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/models/TextGuardReport.java
@@ -1,25 +1,22 @@
-package cloud.pangeacyber.pangea.data_guard.models;
+package cloud.pangeacyber.pangea.ai_guard.models;
 
+import cloud.pangeacyber.pangea.intel.models.UserBreachedData;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import java.util.List;
 import lombok.Value;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(Include.NON_NULL)
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 @Value
-public final class IntelResults {
+public final class TextGuardReport {
 
-	/** The categories that apply to this indicator as determined by the provider. */
-	List<String> category;
-
-	/** The score, given by the Pangea service, for the indicator. */
-	double score;
-
-	/** The verdict for the indicator. */
-	String verdict;
+	IntelResults domainIntel;
+	IntelResults ipIntel;
+	RedactReport redact;
+	IntelResults urlIntel;
+	UserBreachedData userIntel;
 }
