+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/models/TextGuardSecurityIssues.java
@@ -1,20 +1,23 @@
-package cloud.pangeacyber.pangea.data_guard.models;
+package cloud.pangeacyber.pangea.ai_guard.models;
 
-import cloud.pangeacyber.pangea.redact.models.RedactRecognizerResult;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import java.util.List;
-import lombok.Value;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
+@Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(Include.NON_NULL)
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-@Value
-public final class RedactReport {
+@NoArgsConstructor
+public final class TextGuardSecurityIssues {
 
-	int count;
-	List<RedactRecognizerResult> recognizerResults;
+	int compromisedEmailAddresses;
+	int maliciousDomainCount;
+	int maliciousIpCount;
+	int maliciousUrlCount;
+	int redactRuleMatchCount;
 }
