+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/requests/TextGuardRequest.java
@@ -1,4 +1,4 @@
-package cloud.pangeacyber.pangea.data_guard.requests;
+package cloud.pangeacyber.pangea.ai_guard.requests;
 
 import cloud.pangeacyber.pangea.BaseRequest;
 import com.fasterxml.jackson.annotation.JsonInclude;
