+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/results/TextGuardResult.java
@@ -1,8 +1,8 @@
-package cloud.pangeacyber.pangea.data_guard.results;
+package cloud.pangeacyber.pangea.ai_guard.results;
 
-import cloud.pangeacyber.pangea.data_guard.models.TextGuardArtifact;
-import cloud.pangeacyber.pangea.data_guard.models.TextGuardFindings;
-import cloud.pangeacyber.pangea.data_guard.models.TextGuardReport;
+import cloud.pangeacyber.pangea.ai_guard.models.TextGuardArtifact;
+import cloud.pangeacyber.pangea.ai_guard.models.TextGuardFindings;
+import cloud.pangeacyber.pangea.ai_guard.models.TextGuardReport;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
