--- packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/data_guard/responses/TextGuardResponse.java
@@ -1,6 +1,0 @@
-package cloud.pangeacyber.pangea.data_guard.responses;
-
-import cloud.pangeacyber.pangea.Response;
-import cloud.pangeacyber.pangea.data_guard.results.TextGuardResult;
-
-public final class TextGuardResponse extends Response<TextGuardResult> {}
